requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
sqlalchemy
python-dotenv
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
slowapi
python-multipart
midtransclient
psycopg2-binary
bcrypt
passlib
google-auth
google-auth-oauthlib
certifi
email-validator
